requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
jalali_core==1.0.0
jdatetime==5.2.0
mysql-connector-python==9.4.0
pyTelegramBotAPI==4.29.1
requests==2.32.5
urllib3==2.5.0